Traps have gay Chests
trapPopulate() now checks for  a hole two deep or more and places a trap
- need to check for open spaces below placement, may end up janky when Traps are actually implemented

chestPopulate() now implemented. Iterates through x/z's while calling: groundPocketPopulate()

groundPocketPopulate takes in an x,y,z (i,k,j) a replacement (0-Void, 1-Wall, 2-Trap?), and an isolation int (how many spaces back does it have to be on the grid (with only 1 opening) to be considered for a replacement. If it is, randomrange is called
- change  random range on pocket populate to (Random.Range(0,100) < (mapCull  / 20))