Billboarding
Billboarding works in terms of rotation. Moving around the radius of the capsule is another story. Code that looked cool is saved in comments, but code that works  is another story. Another story is that fuck this it's dumb and I gotta go to work.

Check Billboard Scripts when you return to work on this.